♻️ Refactor: 날짜 포맷 함수 utils로 분리
